+++ Packages/com.radtad2.sceneservice/CHANGELOG.md
@@ -1,3 +1,7 @@
-## [1.0.0] - 2025-11-25s
+## [1.1.0] - 2025-12-18
+### Scene Events
+- Added scene events to controller for load start, progressed, and completed.
+
+## [1.0.0] - 2025-11-25
 ### First Release
 - A generic scene transitioner that unifies the scene loading API across multiple providers like Unity and a networking library.